Moving the ConfigureAwait(false) analyzer to ApiDesignGuidelines as an IOperation analyzer and the fixer as a SyntaxGenerator fixer.
