+++ src/Microsoft.ApiDesignGuidelines.Analyzers/UnitTests/Microsoft.ApiDesignGuidelines.Analyzers.UnitTests.csproj
@@ -145,6 +145,8 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DoNotDirectlyAwaitATaskTests.cs" />
+    <Compile Include="DoNotDirectlyAwaitATaskTests.Fixer.cs" />
     <Compile Include="EnumWithFlagsAttributeTests.cs" />
     <Compile Include="EnumWithFlagsAttributeTests.Fixer.cs" />
     <Compile Include="EquatableAnalyzerTests.cs" />
