+++ src/System.Threading.Tasks.Analyzers/CSharp/System.Threading.Tasks.CSharp.Analyzers.csproj
@@ -97,8 +97,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="CSharpDoNotDirectlyAwaitATask.cs" />
-    <Compile Include="CSharpDoNotDirectlyAwaitATask.Fixer.cs" />
     <Compile Include="CSharpDoNotCreateTasksWithoutPassingATaskScheduler.cs" />
     <Compile Include="CSharpDoNotCreateTasksWithoutPassingATaskScheduler.Fixer.cs" />
   </ItemGroup>
