+++ src/System.Threading.Tasks.Analyzers/Core/SystemThreadingTasksAnalyzersResources.resx
@@ -117,16 +117,6 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-
-  <data name="DoNotDirectlyAwaitATaskTitle" xml:space="preserve">
-    <value>Do not directly await a Task</value>
-  </data>
-  <data name="DoNotDirectlyAwaitATaskDescription" xml:space="preserve">
-    <value>#N/A</value>
-  </data>
-  <data name="DoNotDirectlyAwaitATaskMessage" xml:space="preserve">
-    <value>Do not directly await a Task</value>
-  </data>
   <data name="DoNotCreateTasksWithoutPassingATaskSchedulerTitle" xml:space="preserve">
     <value>Do not create tasks without passing a TaskScheduler</value>
   </data>
