+++ src/System.Threading.Tasks.Analyzers/UnitTests/System.Threading.Tasks.Analyzers.UnitTests.csproj
@@ -146,8 +146,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="DoNotDirectlyAwaitATaskTests.cs" />
-    <Compile Include="DoNotDirectlyAwaitATaskTests.Fixer.cs" />
     <Compile Include="DoNotCreateTasksWithoutPassingATaskSchedulerTests.cs" />
     <Compile Include="DoNotCreateTasksWithoutPassingATaskSchedulerTests.Fixer.cs" />
   </ItemGroup>
