--- src/System.Threading.Tasks.Analyzers/VisualBasic/BasicDoNotDirectlyAwaitATask.Fixer.vb
@@ -1,19 +1,0 @@
-' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-Imports System.Composition
-Imports System.Linq
-Imports Microsoft.CodeAnalysis
-Imports Microsoft.CodeAnalysis.CodeFixes
-Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
-
-Namespace System.Threading.Tasks.Analyzers     
-    ''' <summary>
-    ''' RS0003: Do not directly await a Task
-    ''' </summary>
-    <ExportCodeFixProvider(LanguageNames.VisualBasic), [Shared]>
-    Public NotInheritable Class BasicDoNotDirectlyAwaitATaskFixer
-        Inherits DoNotDirectlyAwaitATaskFixer 
-
-    End Class
-End Namespace
