--- src/Unfactored/Roslyn/Core/Reliability/DirectlyAwaitingTaskAnalyzerRule.cs
@@ -1,21 +1,0 @@
-﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using Microsoft.CodeAnalysis;
-
-namespace Roslyn.Diagnostics.Analyzers
-{
-    internal static class DirectlyAwaitingTaskAnalyzerRule
-    {
-        private static readonly LocalizableString s_localizableTitle = new LocalizableResourceString(nameof(RoslynDiagnosticsResources.DirectlyAwaitingTaskDescription), RoslynDiagnosticsResources.ResourceManager, typeof(RoslynDiagnosticsResources));
-        private static readonly LocalizableString s_localizableMessage = new LocalizableResourceString(nameof(RoslynDiagnosticsResources.DirectlyAwaitingTaskMessage), RoslynDiagnosticsResources.ResourceManager, typeof(RoslynDiagnosticsResources));
-
-        public static DiagnosticDescriptor Rule = new DiagnosticDescriptor(
-            RoslynDiagnosticIds.DirectlyAwaitingTaskAnalyzerRuleId,
-            s_localizableTitle,
-            s_localizableMessage,
-            "Reliability",
-            DiagnosticSeverity.Warning,
-            isEnabledByDefault: true,
-            customTags: WellKnownDiagnosticTags.Telemetry);
-    }
-}
